# src/modules/bookmarks/dto/bookmark-create.dto.ts
import { ApiProperty } from "@nestjs/swagger";
import { IsNotEmpty, IsString } from "class-validator";

export class BookmarkCreateDto {
  @ApiProperty()
  @IsString()
  @IsNotEmpty()
  title: string;
}