# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistryDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}}},"tags":["Авторизация"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}}},"tags":["Авторизация"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}}},"tags":["Авторизация"]}},"/api/v1/user":{"get":{"operationId":"UserController_self","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["Пользователи"]}},"/api/v1/bookmarks":{"get":{"operationId":"BookmarksController_getBookmarks","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookmarkDto"}}}}}},"tags":["Закладки"]},"post":{"operationId":"BookmarksController_createBookmark","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookmarkCreateDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookmarkDto"}}}}},"tags":["Закладки"]}},"/api/v1/bookmarks/{id}":{"patch":{"operationId":"BookmarksController_updateBookmark","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookmarkCreateDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookmarkDto"}}}}},"tags":["Закладки"]},"delete":{"operationId":"BookmarksController_deleteBookmark","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookmarkDto"}}}}},"tags":["Закладки"]}},"/api/v1/bookmarks/recipes":{"get":{"operationId":"BookmarksController_getRecipesInBookmarks","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookmarkRecipeDto"}}}}}},"tags":["Закладки"]}},"/api/v1/bookmarks/{bookmarkId}/recipes/{recipeId}":{"get":{"operationId":"BookmarksController_addRecipeToBookmark","parameters":[{"name":"recipeId","required":true,"in":"path","schema":{"type":"string"}},{"name":"bookmarkId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Закладки"]}},"/api/v1/bookmarks/recipes/{recipeId}":{"delete":{"operationId":"BookmarksController_removeRecipeFromBookmark","parameters":[{"name":"recipeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Закладки"]}},"/api/v1/recipes/search":{"post":{"operationId":"RecipeController_getRecipes","parameters":[],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipesFilterDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RecipeDto"}}}}}},"tags":["Рецепты"]}},"/api/v1/recipes/slug/{slug}":{"get":{"operationId":"RecipeController_getRecipeBySlug","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDto"}}}}},"tags":["Рецепты"]}},"/api/v1/recipes/{id}":{"get":{"operationId":"RecipeController_getRecipeById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDto"}}}}},"tags":["Рецепты"]}},"/api/v1/recipes":{"post":{"operationId":"RecipeController_saveRecipe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeCreateDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDto"}}}}},"tags":["Рецепты"]}},"/api/v1/recipes/{slug}":{"put":{"operationId":"RecipeController_updateRecipe","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDto"}}}},"responses":{"200":{"description":""}},"tags":["Рецепты"]},"delete":{"operationId":"RecipeController_deleteRecipe","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Рецепты"]}},"/api/v1/ingredients":{"get":{"operationId":"RecipeIngredientController_getRecipeIngredients","parameters":[{"name":"q","required":true,"in":"query","description":"Строка для поиска","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IngredientDto"}}}}}},"tags":["Ингредиенты"]}},"/api/v1/ingredients/amount-types":{"get":{"operationId":"RecipeIngredientController_getTypes","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AmountTypeDto"}}}}}},"tags":["Ингредиенты"]}},"/api/v1/ingredients/{id}":{"get":{"operationId":"RecipeIngredientController_getRecipeIngredientById","parameters":[{"name":"id","required":true,"in":"path","schema":{}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngredientDto"}}}}},"tags":["Ингредиенты"]}}},"info":{"title":"Рецепты","description":"Апи для сайта рецептов","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UserRegistryDto":{"type":"object","properties":{"userName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","minLength":3}},"required":["userName","email","password"]},"TokenResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"UserLoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"userName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","RECIPES_MODERATOR","COMMENTS_MODERATOR","USER"]},"avatar":{"type":"string"},"createdAt":{"type":"string"},"updateAt":{"type":"string"}},"required":["id","userName","email","role","avatar","createdAt","updateAt"]},"BookmarkDto":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"slug":{"type":"string"},"createdAt":{"type":"string"},"updateAt":{"type":"string"}},"required":["id","title","slug","createdAt","updateAt"]},"BookmarkCreateDto":{"type":"object","properties":{"title":{"type":"string"}},"required":["title"]},"BookmarkRecipeDto":{"type":"object","properties":{"id":{"type":"string"},"recipeId":{"type":"string"},"bookmarkId":{"type":"string"},"createdAt":{"type":"string"}},"required":["id","recipeId","bookmarkId","createdAt"]},"IngredientsFilterDto":{"type":"object","properties":{"includes":{"type":"array","items":{"type":"string"}},"excludes":{"type":"array","items":{"type":"string"}}}},"RecipesFilterDto":{"type":"object","properties":{"q":{"type":"string","description":"Строка поиска по названиям рецептов"},"slugs":{"type":"array","items":{"type":"string"}},"ingredients":{"description":"Списки id ингредиентов","allOf":[{"$ref":"#/components/schemas/IngredientsFilterDto"}]},"isDeleted":{"type":"boolean","description":"Показать рецепты помеченные удаленными (только для админов!)"}}},"RecipeIngredientDto":{"type":"object","properties":{"id":{"type":"string"},"count":{"type":"number"},"ingredientId":{"type":"number"},"amountTypeId":{"type":"number"},"createdAt":{"type":"string"}},"required":["id","count","ingredientId","amountTypeId","createdAt"]},"RecipeDto":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"slug":{"type":"string"},"description":{"type":"string"},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/RecipeIngredientDto"}},"images":{"type":"array","items":{"type":"string"}},"steps":{"type":"array","items":{"type":"array"}},"user":{"$ref":"#/components/schemas/UserDto"},"isDeleted":{"type":"boolean"},"createdAt":{"type":"string"},"updateAt":{"type":"string"}},"required":["id","title","slug","description","ingredients","images","steps","user","isDeleted","createdAt","updateAt"]},"RecipeCreateDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"images":{"type":"array","items":{"type":"string"}},"steps":{"type":"array","items":{"type":"string"}},"ingredients":{"type":"array","items":{"type":"string"}}},"required":["title","description","tags","images","steps","ingredients"]},"IngredientDto":{"type":"object","properties":{"id":{"type":"number"},"slug":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"amountTypes":{"type":"array","items":{"type":"number"}},"createdAt":{"type":"string"},"updateAt":{"type":"string"}},"required":["id","slug","name","description","amountTypes","createdAt","updateAt"]},"AmountTypeDto":{"type":"object","properties":{"id":{"type":"number"},"slug":{"type":"string"},"name":{"type":"string"},"createdAt":{"type":"string"},"updateAt":{"type":"string"}},"required":["id","slug","name","createdAt","updateAt"]}}}}